requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.6.0
python-multipart==0.0.9
pypdf==4.3.1
chromadb==0.5.3
openai==1.59.3
google-generativeai==0.8.3

# NEW for doc types + diagrams
python-docx==1.1.2
openpyxl==3.1.5
pymupdf==1.24.10
